staff init and calim implemented

+++ programs/anchor-payroll-capstone-q1-26/src/state/staff.rs
@@ -4,7 +4,7 @@ use anchor_lang::prelude::*;
 #[derive(InitSpace)]
 pub struct StaffAccount {
     pub active: bool,
-    pub rate: u8,
+    pub rate: u64,
     pub total_claimed: u64,
     pub time_started: u64
 }
